+++ jsonschema/src/resolver.rs
@@ -11,8 +11,111 @@ use serde_json::Value;
 use std::{borrow::Cow, sync::Arc};
 use url::Url;
 
-#[derive(Debug)]
+/// An opaque error type that is returned
+/// by resolvers on resolution failures.
+pub type SchemaResolverError = anyhow::Error;
+
+/// A resolver that resolves external schema references.
+/// Internal references such as `#/definitions` and JSON pointers are handled internally.
+///
+/// All operations are blocking and it is not possible to return futures.
+/// As a workaround, errors can be returned that will contain the schema URLs to resolve
+/// and can be resolved outside the validation process if needed.
+///
+/// # Example
+///
+/// ```no_run
+/// # use serde_json::{json, Value};
+/// # use anyhow::anyhow;
+/// # use jsonschema::{SchemaResolver, SchemaResolverError};
+/// # use std::sync::Arc;
+/// # use url::Url;
+///
+/// struct MyCustomResolver;
+///
+/// impl SchemaResolver for MyCustomResolver {
+///     fn resolve(&self, root_schema: &Value, url: &Url, _original_reference: &str) -> Result<Arc<Value>, SchemaResolverError> {
+///         match url.scheme() {
+///             "json-schema" => {
+///                 Err(anyhow!("cannot resolve schema without root schema ID"))
+///             },
+///             "http" | "https" => {
+///                 Ok(Arc::new(json!({ "description": "an external schema" })))
+///             }
+///             _ => Err(anyhow!("scheme is not supported"))
+///         }
+///     }
+/// }
+/// ```
+pub trait SchemaResolver: Send + Sync {
+    /// Resolve an external schema via an URL.
+    ///
+    /// Relative URLs are resolved based on the root schema's ID,
+    /// if there is no root schema ID available, the scheme `json-schema` is used
+    /// and any relative paths are turned into absolutes.
+    ///
+    /// Additionally the original reference string is also passed,
+    /// in most cases it should not be needed, but it preserves some information,
+    /// such as relative paths that are lost when the URL is built.
+    fn resolve(
+        &self,
+        root_schema: &Value,
+        url: &Url,
+        original_reference: &str,
+    ) -> Result<Arc<Value>, SchemaResolverError>;
+}
+
+pub(crate) struct DefaultResolver;
+
+impl SchemaResolver for DefaultResolver {
+    fn resolve(
+        &self,
+        _root_schema: &Value,
+        url: &Url,
+        _reference: &str,
+    ) -> Result<Arc<Value>, SchemaResolverError> {
+        match url.scheme() {
+            "http" | "https" => {
+                #[cfg(all(feature = "reqwest", not(feature = "resolve-http")))]
+                {
+                    compile_error!("the `reqwest` feature does not enable HTTP schema resolving anymore, use the `resolve-http` feature instead");
+                }
+
+                #[cfg(any(feature = "resolve-http", test))]
+                {
+                    let response = reqwest::blocking::get(url.as_str())?;
+                    let document: Value = response.json()?;
+                    Ok(Arc::new(document))
+                }
+                #[cfg(not(any(feature = "resolve-http", test)))]
+                Err(anyhow::anyhow!("`resolve-http` feature or a custom resolver is required to resolve external schemas via HTTP"))
+            }
+            "file" => {
+                #[cfg(any(feature = "resolve-file", test))]
+                {
+                    if let Ok(path) = url.to_file_path() {
+                        let f = std::fs::File::open(path)?;
+                        let document: Value = serde_json::from_reader(f)?;
+                        Ok(Arc::new(document))
+                    } else {
+                        Err(anyhow::anyhow!("invalid file path"))
+                    }
+                }
+                #[cfg(not(any(feature = "resolve-file", test)))]
+                {
+                    Err(anyhow::anyhow!("`resolve-file` feature or a custom resolver is required to resolve external schemas via files"))
+                }
+            }
+            "json-schema" => Err(anyhow::anyhow!(
+                "cannot resolve relative external schema without root schema ID"
+            )),
+            _ => Err(anyhow::anyhow!("unknown scheme {}", url.scheme())),
+        }
+    }
+}
+
 pub(crate) struct Resolver {
+    external_resolver: Arc<dyn SchemaResolver>,
     root_schema: Arc<Value>,
     // canonical_id: sub-schema mapping to resolve documents by their ID
     // canonical_id is composed with the root document id
@@ -21,8 +124,19 @@ pub(crate) struct Resolver {
     store: RwLock<AHashMap<String, Arc<Value>>>,
 }
 
+impl std::fmt::Debug for Resolver {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("Resolver")
+            .field("root_schema", &self.root_schema)
+            .field("schemas", &self.schemas)
+            .field("store", &self.store)
+            .finish()
+    }
+}
+
 impl Resolver {
     pub(crate) fn new<'a>(
+        external_resolver: Arc<dyn SchemaResolver>,
         draft: Draft,
         scope: &Url,
         schema: Arc<Value>,
@@ -35,6 +149,7 @@ impl Resolver {
             None
         })?;
         Ok(Resolver {
+            external_resolver,
             root_schema: schema,
             schemas,
             store: RwLock::new(store),
@@ -46,36 +161,25 @@ impl Resolver {
     ///   - the root document (`DEFAULT_ROOT_URL`) case;
     ///   - named subschema that is stored in `self.schemas`;
     ///   - document from a remote location;
-    fn resolve_url<'a>(&'a self, url: &Url) -> Result<Arc<Value>, ValidationError> {
+    fn resolve_url<'a>(&'a self, url: &Url, orig_ref: &str) -> Result<Arc<Value>, ValidationError> {
         match url.as_str() {
             DEFAULT_ROOT_URL => Ok(self.root_schema.clone()),
-            url_str => {
-                if let Some(cached) = self.store.read().get(url_str) {
-                    return Ok(cached.clone());
+            url_str => match self.schemas.get(url_str) {
+                Some(value) => Ok(value.clone()),
+                None => {
+                    if let Some(cached) = self.store.read().get(url_str) {
+                        return Ok(cached.clone());
+                    }
+                    let resolved = self
+                        .external_resolver
+                        .resolve(&self.root_schema, url, orig_ref)
+                        .map_err(|error| ValidationError::resolver(url.clone(), error))?;
+                    self.store
+                        .write()
+                        .insert(url.clone().into(), resolved.clone());
+                    Ok(resolved)
                 }
-                match self.schemas.get(url_str) {
-                    Some(value) => Ok(value.clone()),
-                    None => match url.scheme() {
-                        "http" | "https" => {
-                            #[cfg(any(feature = "reqwest", test))]
-                            {
-                                let response = reqwest::blocking::get(url.as_str())?;
-                                let document: Value = response.json()?;
-                                let arcdoc = Arc::new(document);
-                                self.store
-                                    .write()
-                                    .insert(url_str.to_string(), arcdoc.clone());
-                                Ok(arcdoc)
-                            }
-                            #[cfg(not(any(feature = "reqwest", test)))]
-                            panic!("trying to resolve an http(s), but reqwest support has not been included");
-                        }
-                        http_scheme => Err(ValidationError::unknown_reference_scheme(
-                            http_scheme.to_owned(),
-                        )),
-                    },
-                }
-            }
+            },
         }
     }
 
@@ -88,6 +192,7 @@ impl Resolver {
         &self,
         draft: Draft,
         url: &Url,
+        orig_ref: &str,
     ) -> Result<(Url, Arc<Value>), ValidationError> {
         let mut resource = url.clone();
         resource.set_fragment(None);
@@ -102,7 +207,7 @@ impl Resolver {
 
         // Each resolved document may be in a changed subfolder
         // They are tracked when JSON pointer is resolved and added to the resource
-        let document = self.resolve_url(&resource)?;
+        let document = self.resolve_url(&resource, orig_ref)?;
         if fragment.is_empty() {
             return Ok((resource, Arc::clone(&document)));
         }
@@ -259,6 +364,7 @@ mod tests {
 
     fn make_resolver(schema: &Value) -> Resolver {
         Resolver::new(
+            Arc::new(DefaultResolver),
             Draft::Draft7,
             &Url::parse("json-schema:///").unwrap(),
             Arc::new(schema.clone()),
@@ -508,7 +614,9 @@ mod tests {
         });
         let resolver = make_resolver(&schema);
         let url = Url::parse("json-schema:///#/definitions/a").unwrap();
-        let (resource, resolved) = resolver.resolve_fragment(Draft::Draft7, &url).unwrap();
+        let (resource, resolved) = resolver
+            .resolve_fragment(Draft::Draft7, &url, "#/definitions/a")
+            .unwrap();
         assert_eq!(resource, Url::parse("json-schema:///").unwrap());
         assert_eq!(resolved.as_ref(), schema.pointer("/definitions/a").unwrap());
     }