+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## [Unreleased]
 
+### Added
+
+- The `SchemaResolver` trait to support resolving external schema references. [#341](https://github.com/Stranger6667/jsonschema-rs/pull/341)
+- `resolve-file` feature to resolve external schema files via `std::fs`. [#341](https://github.com/Stranger6667/jsonschema-rs/pull/341)
+
+### Changed
+
+- The `reqwest` feature was changed to `resolve-http`. [#341](https://github.com/Stranger6667/jsonschema-rs/pull/341)
+
 ### Performance
 
 - CLI: Use `serde::from_reader` instead of `serde::from_str`.